backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-environ==0.11.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
requests==2.32.3
resend==2.0.0
gunicorn==22.0.0
